Style suggestions for permutation tests.
